+++ wallarm/resource_rule_bola_counter_test.go
@@ -2,6 +2,7 @@ package wallarm
 
 import (
 	"fmt"
+	"os"
 	"regexp"
 	"strconv"
 	"testing"
@@ -13,6 +14,9 @@ import (
 )
 
 func TestAccRuleBolaCounterCreate(t *testing.T) {
+	if os.Getenv("WALLARM_EXTRA_PERMISSIONS") == "" {
+		t.Skip("Skipping not test as it requires WALLARM_EXTRA_PERMISSIONS set")
+	}
 	rnd := generateRandomResourceName(5)
 	name := "wallarm_rule_bola_counter." + rnd
 	resource.Test(t, resource.TestCase{
@@ -24,7 +28,7 @@ func TestAccRuleBolaCounterCreate(t *testing.T) {
 				Config: testAccRuleBolaCounterCreate(rnd),
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttr(name, "action.#", "3"),
-					resource.TestMatchResourceAttr(name, "counter", regexp.MustCompile("^d:.+")),
+					resource.TestMatchResourceAttr(name, "counter", regexp.MustCompile("^i:.+")),
 				),
 				ExpectNonEmptyPlan: true,
 			},