+++ wallarm/resource_rule_overlimit_res_settings.go
@@ -13,13 +13,13 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/helper/validation"
 )
 
-func resourceWallarmIgnoreRegex() *schema.Resource {
+func resourceWallarmOverlimitResSettings() *schema.Resource {
 	return &schema.Resource{
-		Create: resourceWallarmIgnoreRegexCreate,
-		Read:   resourceWallarmIgnoreRegexRead,
-		Delete: resourceWallarmIgnoreRegexDelete,
+		Create: resourceWallarmOverlimitResSettingsCreate,
+		Read:   resourceWallarmOverlimitResSettingsRead,
+		Delete: resourceWallarmOverlimitResSettingsDelete,
 		Importer: &schema.ResourceImporter{
-			State: resourceWallarmIgnoreRegexImport,
+			State: resourceWallarmOverlimitResSettingsImport,
 		},
 
 		Schema: map[string]*schema.Schema{
@@ -59,12 +59,6 @@ func resourceWallarmIgnoreRegex() *schema.Resource {
 				ForceNew: true,
 			},
 
-			"regex_id": {
-				Type:     schema.TypeInt,
-				Required: true,
-				ForceNew: true,
-			},
-
 			"action": {
 				Type:     schema.TypeSet,
 				Optional: true,
@@ -95,6 +89,7 @@ func resourceWallarmIgnoreRegex() *schema.Resource {
 										Type:     schema.TypeList,
 										Optional: true,
 										ForceNew: true,
+										Computed: true,
 										Elem:     &schema.Schema{Type: schema.TypeString},
 									},
 									"method": {
@@ -122,31 +117,35 @@ func resourceWallarmIgnoreRegex() *schema.Resource {
 										Type:     schema.TypeString,
 										Optional: true,
 										ForceNew: true,
+										Computed: true,
 									},
 
 									"action_ext": {
 										Type:     schema.TypeString,
 										Optional: true,
 										ForceNew: true,
+										Computed: true,
 									},
 
 									"query": {
 										Type:     schema.TypeString,
 										Optional: true,
 										ForceNew: true,
+										Computed: true,
 									},
 
 									"proto": {
-										Type:         schema.TypeString,
-										Optional:     true,
-										ForceNew:     true,
-										ValidateFunc: validation.StringInSlice([]string{"1.0", "1.1", "2.0", "3.0"}, false),
+										Type:     schema.TypeString,
+										Optional: true,
+										ForceNew: true,
+										Computed: true,
 									},
 
 									"scheme": {
 										Type:         schema.TypeString,
 										Optional:     true,
 										ForceNew:     true,
+										Computed:     true,
 										ValidateFunc: validation.StringInSlice([]string{"http", "https"}, true),
 									},
 
@@ -154,6 +153,7 @@ func resourceWallarmIgnoreRegex() *schema.Resource {
 										Type:     schema.TypeString,
 										Optional: true,
 										ForceNew: true,
+										Computed: true,
 									},
 
 									"instance": {
@@ -163,7 +163,7 @@ func resourceWallarmIgnoreRegex() *schema.Resource {
 										ValidateFunc: func(val interface{}, key string) (warns []string, errs []error) {
 											v := val.(int)
 											if v < -1 {
-												errs = append(errs, fmt.Errorf("%q must be be greater than -1 inclusive, got: %d", key, v))
+												errs = append(errs, fmt.Errorf("%q must be greater than -1 inclusive, got: %d", key, v))
 											}
 											return
 										},
@@ -177,34 +177,35 @@ func resourceWallarmIgnoreRegex() *schema.Resource {
 
 			"point": {
 				Type:     schema.TypeList,
-				Required: true,
+				Optional: true,
 				ForceNew: true,
 				Elem: &schema.Schema{
 					Type: schema.TypeList,
-					Elem: &schema.Schema{
-						Type: schema.TypeString,
-					},
+					Elem: &schema.Schema{Type: schema.TypeString},
 				},
+			},
+
+			"overlimit_time": {
+				Type:         schema.TypeInt,
+				ForceNew:     true,
+				Required:     true,
+				ValidateFunc: validation.IntBetween(0, 2_147_483_647),
+			},
+
+			"mode": {
+				Type:         schema.TypeString,
+				Required:     true,
+				ForceNew:     true,
+				ValidateFunc: validation.StringInSlice([]string{"off", "monitoring", "blocking"}, false),
 			},
 		},
 	}
 }
 
-func resourceWallarmIgnoreRegexCreate(d *schema.ResourceData, m interface{}) error {
+func resourceWallarmOverlimitResSettingsCreate(d *schema.ResourceData, m interface{}) error {
 	client := m.(wallarm.API)
 	clientID := retrieveClientID(d, client)
 	comment := d.Get("comment").(string)
-	regexID := d.Get("regex_id").(int)
-
-	ps := d.Get("point").([]interface{})
-	if err := d.Set("point", ps); err != nil {
-		return err
-	}
-
-	points, err := expandPointsToTwoDimensionalArray(ps)
-	if err != nil {
-		return err
-	}
 
 	actionsFromState := d.Get("action").(*schema.Set)
 	action, err := expandSetToActionDetailsList(actionsFromState)
@@ -212,61 +213,54 @@ func resourceWallarmIgnoreRegexCreate(d *schema.ResourceData, m interface{}) err
 		return err
 	}
 
-	vp := &wallarm.ActionCreate{
-		Type:                "disable_regex",
-		Clientid:            clientID,
-		Action:              &action,
-		Point:               points,
-		RegexID:             regexID,
-		Comment:             comment,
-		VariativityDisabled: true,
-	}
-
-	actionResp, err := client.HintCreate(vp)
+	iPoint := d.Get("point").([]interface{})
+	point, err := expandPointsToTwoDimensionalArray(iPoint)
 	if err != nil {
 		return err
 	}
+	overlimitTime := d.Get("overlimit_time").(int)
+	mode := d.Get("mode").(string)
+
+	actionBody := &wallarm.ActionCreate{
+		Type:          "overlimit_res_settings",
+		Clientid:      clientID,
+		Action:        &action,
+		Validated:     false,
+		Comment:       comment,
+		Point:         point,
+		Mode:          mode,
+		OverlimitTime: overlimitTime,
+	}
+
+	actionResp, err := client.HintCreate(actionBody)
+	if err != nil {
+		return err
+	}
+	actionID := actionResp.Body.ActionID
 
 	d.Set("rule_id", actionResp.Body.ID)
-	d.Set("action_id", actionResp.Body.ActionID)
+	d.Set("action_id", actionID)
 	d.Set("rule_type", actionResp.Body.Type)
+	d.Set("client_id", clientID)
+	d.Set("point", actionResp.Body.Point)
 
-	resID := fmt.Sprintf("%d/%d/%d/%s", clientID, actionResp.Body.ActionID, actionResp.Body.ID, actionResp.Body.Type)
+	resID := fmt.Sprintf("%d/%d/%d", clientID, actionID, actionResp.Body.ID)
 	d.SetId(resID)
 
-	return resourceWallarmIgnoreRegexRead(d, m)
+	return nil
 }
 
-func resourceWallarmIgnoreRegexRead(d *schema.ResourceData, m interface{}) error {
+func resourceWallarmOverlimitResSettingsRead(d *schema.ResourceData, m interface{}) error {
 	client := m.(wallarm.API)
 	clientID := retrieveClientID(d, client)
 	actionID := d.Get("action_id").(int)
 	ruleID := d.Get("rule_id").(int)
-	regexID := d.Get("regex_id").(int)
-
-	ps := d.Get("point").([]interface{})
-	var points []interface{}
-	for _, point := range ps {
-		p := point.([]interface{})
-		points = append(points, p...)
-	}
 
 	actionsFromState := d.Get("action").(*schema.Set)
 	action, err := expandSetToActionDetailsList(actionsFromState)
 	if err != nil {
 		return err
 	}
-
-	var actsSlice []interface{}
-	for _, a := range action {
-		acts, err := actionDetailsToMap(a)
-		if err != nil {
-			return err
-		}
-		actsSlice = append(actsSlice, acts)
-	}
-
-	actionsSet := schema.NewSet(hashResponseActionDetails, actsSlice)
 
 	hint := &wallarm.HintRead{
 		Limit:     1000,
@@ -289,9 +283,8 @@ func resourceWallarmIgnoreRegexRead(d *schema.ResourceData, m interface{}) error
 
 	expectedRule := wallarm.ActionBody{
 		ActionID: actionID,
-		Type:     "disable_regex",
-		Point:    points,
-		RegexID:  regexID,
+		Type:     "overlimit_res_settings",
+		Action:   action,
 	}
 
 	var notFoundRules []int
@@ -302,15 +295,10 @@ func resourceWallarmIgnoreRegexRead(d *schema.ResourceData, m interface{}) error
 			continue
 		}
 
-		// The response has a different structure so we have to align them
-		// to uniform view then to compare.
-		alignedPoints := alignPointScheme(rule.Point)
-
 		actualRule := &wallarm.ActionBody{
 			ActionID: rule.ActionID,
 			Type:     rule.Type,
-			Point:    alignedPoints,
-			RegexID:  rule.RegexID,
+			Action:   rule.Action,
 		}
 
 		if cmp.Equal(expectedRule, *actualRule) && equalWithoutOrder(action, rule.Action) {
@@ -327,14 +315,6 @@ func resourceWallarmIgnoreRegexRead(d *schema.ResourceData, m interface{}) error
 
 	d.Set("client_id", clientID)
 
-	if actionsSet.Len() != 0 {
-		if err := d.Set("action", &actionsSet); err != nil {
-			return err
-		}
-	} else {
-		log.Printf("[WARN] action was empty so it either doesn't exist or it is a default branch which has no conditions. Actions: %v", &actionsSet)
-	}
-
 	if updatedRuleID == 0 {
 		log.Printf("[WARN] these rule IDs: %v have been found under the action ID: %d. But it isn't in the Terraform Plan.", notFoundRules, actionID)
 		d.SetId("")
@@ -343,47 +323,26 @@ func resourceWallarmIgnoreRegexRead(d *schema.ResourceData, m interface{}) error
 	return nil
 }
 
-func resourceWallarmIgnoreRegexDelete(d *schema.ResourceData, m interface{}) error {
+func resourceWallarmOverlimitResSettingsDelete(d *schema.ResourceData, m interface{}) error {
 	client := m.(wallarm.API)
 	clientID := retrieveClientID(d, client)
-	actionID := d.Get("action_id").(int)
+	ruleID := d.Get("rule_id").(int)
 
-	rule := &wallarm.ActionRead{
-		Filter: &wallarm.ActionFilter{
-			HintType: []string{"disable_regex"},
+	h := &wallarm.HintDelete{
+		Filter: &wallarm.HintDeleteFilter{
 			Clientid: []int{clientID},
-			ID:       []int{actionID},
+			ID:       ruleID,
 		},
-		Limit:  1000,
-		Offset: 0,
 	}
-	respRules, err := client.RuleRead(rule)
-	if err != nil {
+
+	if err := client.HintDelete(h); err != nil {
 		return err
-	}
-
-	if len(respRules.Body) == 1 && respRules.Body[0].Hints == 1 && respRules.Body[0].GroupedHintsCount == 1 {
-		if err := client.RuleDelete(actionID); err != nil {
-			return err
-		}
-	} else {
-		ruleID := d.Get("rule_id").(int)
-		h := &wallarm.HintDelete{
-			Filter: &wallarm.HintDeleteFilter{
-				Clientid: []int{clientID},
-				ID:       ruleID,
-			},
-		}
-
-		if err := client.HintDelete(h); err != nil {
-			return err
-		}
 	}
 
 	return nil
 }
 
-func resourceWallarmIgnoreRegexImport(d *schema.ResourceData, m interface{}) ([]*schema.ResourceData, error) {
+func resourceWallarmOverlimitResSettingsImport(d *schema.ResourceData, m interface{}) ([]*schema.ResourceData, error) {
 	client := m.(wallarm.API)
 	idAttr := strings.SplitN(d.Id(), "/", 3)
 	if len(idAttr) == 3 {
@@ -401,7 +360,7 @@ func resourceWallarmIgnoreRegexImport(d *schema.ResourceData, m interface{}) ([]
 		}
 		d.Set("action_id", actionID)
 		d.Set("rule_id", ruleID)
-		d.Set("rule_type", "disable_regex")
+		d.Set("rule_type", "overlimit_res_settings")
 
 		hint := &wallarm.HintRead{
 			Limit:     1000,
@@ -411,7 +370,7 @@ func resourceWallarmIgnoreRegexImport(d *schema.ResourceData, m interface{}) ([]
 			Filter: &wallarm.HintFilter{
 				Clientid: []int{clientID},
 				ID:       []int{ruleID},
-				Type:     []string{"disable_regex"},
+				Type:     []string{"overlimit_res_settings"},
 			},
 		}
 		actionHints, err := client.HintRead(hint)
@@ -434,13 +393,11 @@ func resourceWallarmIgnoreRegexImport(d *schema.ResourceData, m interface{}) ([]
 			if err := d.Set("action", &actionsSet); err != nil {
 				return nil, err
 			}
-			d.Set("regex_id", (*actionHints.Body)[0].RegexID)
 
 		}
 
-		pointInterface := (*actionHints.Body)[0].Point
-		point := wrapPointElements(pointInterface)
-		d.Set("point", point)
+		d.Set("mode", (*actionHints.Body)[0].Mode)
+		d.Set("overlimit_time", (*actionHints.Body)[0].OverlimitTime)
 
 		existingID := fmt.Sprintf("%d/%d/%d", clientID, actionID, ruleID)
 		d.SetId(existingID)
@@ -449,7 +406,7 @@ func resourceWallarmIgnoreRegexImport(d *schema.ResourceData, m interface{}) ([]
 		return nil, fmt.Errorf("invalid id (%q) specified, should be in format \"{clientID}/{actionID}/{ruleID}\"", d.Id())
 	}
 
-	resourceWallarmIgnoreRegexRead(d, m)
+	resourceWallarmOverlimitResSettingsRead(d, m)
 
 	return []*schema.ResourceData{d}, nil
 }