+++ wallarm/resource_rule_regex.go
@@ -68,8 +68,6 @@ func resourceWallarmRegex() *schema.Resource {
 				Type:     schema.TypeString,
 				Required: true,
 				ForceNew: true,
-				ValidateFunc: validation.StringInSlice([]string{"any", "sqli", "rce", "crlf", "nosqli", "ptrav",
-					"xxe", "ptrav", "xss", "scanner", "redir", "ldapi", "vpatch"}, false),
 			},
 
 			"action": {
@@ -319,7 +317,7 @@ func resourceWallarmRegexRead(d *schema.ResourceData, m interface{}) error {
 		OrderDesc: true,
 		Filter: &wallarm.HintFilter{
 			Clientid: []int{clientID},
-			ActionID: []int{actionID},
+			ID:       []int{ruleID},
 		},
 	}
 	actionHints, err := client.HintRead(hint)
@@ -466,10 +464,9 @@ func resourceWallarmRegexImport(d *schema.ResourceData, m interface{}) ([]*schem
 			d.Set("regex", (*actionHints.Body)[0].Regex)
 			d.Set("attack_type", (*actionHints.Body)[0].AttackType)
 
-			// TODO: Import point as the struct changes when API responds.
-			// Currently it doesn't import it correctly so it leads to
-			// a necessity to apply changes again via destroy/create function.
-			d.Set("point", (*actionHints.Body)[0].Point)
+			pointInterface := (*actionHints.Body)[0].Point
+			point := wrapPointElements(pointInterface)
+			d.Set("point", point)
 		}
 
 		if hintType == "experimental_regex" {