+++ wallarm/resource_rule_variative_keys_test.go
@@ -2,6 +2,7 @@ package wallarm
 
 import (
 	"fmt"
+	"os"
 	"strconv"
 	"testing"
 
@@ -12,6 +13,9 @@ import (
 )
 
 func TestAccRuleVariativeKeysCreate_Basic(t *testing.T) {
+	if os.Getenv("WALLARM_EXTRA_PERMISSIONS") == "" {
+		t.Skip("Skipping not test as it requires WALLARM_EXTRA_PERMISSIONS set")
+	}
 	rnd := generateRandomResourceName(5)
 	name := "wallarm_rule_variative_keys." + rnd
 	resource.Test(t, resource.TestCase{
@@ -32,6 +36,9 @@ func TestAccRuleVariativeKeysCreate_Basic(t *testing.T) {
 }
 
 func TestAccRuleVariativeKeysCreateRecreate(t *testing.T) {
+	if os.Getenv("WALLARM_EXTRA_PERMISSIONS") == "" {
+		t.Skip("Skipping not test as it requires WALLARM_EXTRA_PERMISSIONS set")
+	}
 	rnd := generateRandomResourceName(5)
 	name := "wallarm_rule_variative_keys." + rnd
 	resource.Test(t, resource.TestCase{