+++ wallarm/resource_rule_variative_values.go
@@ -265,7 +265,7 @@ func resourceWallarmVariativeValuesRead(d *schema.ResourceData, m interface{}) e
 		OrderDesc: true,
 		Filter: &wallarm.HintFilter{
 			Clientid: []int{clientID},
-			ActionID: []int{actionID},
+			ID:       []int{ruleID},
 			Type:     []string{"variative_values"},
 		},
 	}
@@ -424,6 +424,10 @@ func resourceWallarmVariativeValuesImport(d *schema.ResourceData, m interface{})
 				return nil, err
 			}
 		}
+
+		pointInterface := (*actionHints.Body)[0].Point
+		point := wrapPointElements(pointInterface)
+		d.Set("point", point)
 
 		existingID := fmt.Sprintf("%d/%d/%d", clientID, actionID, ruleID)
 		d.SetId(existingID)