+++ wallarm/resource_rule_variative_values_test.go
@@ -2,6 +2,7 @@ package wallarm
 
 import (
 	"fmt"
+	"os"
 	"strconv"
 	"testing"
 
@@ -11,6 +12,9 @@ import (
 )
 
 func TestAccRuleVariativeValuesCreate_Basic(t *testing.T) {
+	if os.Getenv("WALLARM_EXTRA_PERMISSIONS") == "" {
+		t.Skip("Skipping not test as it requires WALLARM_EXTRA_PERMISSIONS set")
+	}
 	rnd := generateRandomResourceName(5)
 	name := "wallarm_rule_variative_values." + rnd
 	resource.Test(t, resource.TestCase{
@@ -31,6 +35,9 @@ func TestAccRuleVariativeValuesCreate_Basic(t *testing.T) {
 }
 
 func TestAccRuleVariativeValuesCreateRecreate(t *testing.T) {
+	if os.Getenv("WALLARM_EXTRA_PERMISSIONS") == "" {
+		t.Skip("Skipping not test as it requires WALLARM_EXTRA_PERMISSIONS set")
+	}
 	rnd := generateRandomResourceName(5)
 	name := "wallarm_rule_variative_values." + rnd
 	resource.Test(t, resource.TestCase{