+++ wallarm/resource_rule_vpatch_test.go
@@ -11,13 +11,6 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/terraform"
 )
 
-type vpatchTestingRule struct {
-	attackType string
-	point      string
-	matchType  []string
-	value      string
-}
-
 func TestAccRuleVpatchCreate_Basic(t *testing.T) {
 	rnd := generateRandomResourceName(5)
 	name := "wallarm_rule_vpatch." + rnd
@@ -30,7 +23,7 @@ func TestAccRuleVpatchCreate_Basic(t *testing.T) {
 			{
 				Config: testWallarmRuleVpatchBasicConfig(rnd, "xss", "iequal", "vpatch.wallarm.com", "HOST", "get_all"),
 				Check: resource.ComposeTestCheckFunc(
-					resource.TestCheckResourceAttr(name, "attack_type.0", "xss"),
+					resource.TestCheckResourceAttr(name, "attack_type", "xss"),
 					resource.TestCheckResourceAttr(name, "action.#", "1"),
 					resource.TestCheckResourceAttr(name, "point.0.0", "get_all"),
 				),
@@ -42,7 +35,6 @@ func TestAccRuleVpatchCreate_Basic(t *testing.T) {
 func TestAccRuleVpatchCreate_DefaultBranch(t *testing.T) {
 	rnd := generateRandomResourceName(5)
 	name := "wallarm_rule_vpatch." + rnd
-	attackType := `"crlf", "scanner", "redir", "ldapi"`
 	point := `["get", "query"]`
 
 	resource.Test(t, resource.TestCase{
@@ -51,12 +43,9 @@ func TestAccRuleVpatchCreate_DefaultBranch(t *testing.T) {
 		CheckDestroy: testAccCheckWallarmRuleVpatchDestroy,
 		Steps: []resource.TestStep{
 			{
-				Config: testWallarmRuleVpatchDefaultBranchConfig(rnd, attackType, point),
+				Config: testWallarmRuleVpatchDefaultBranchConfig(rnd, `"crlf"`, point),
 				Check: resource.ComposeTestCheckFunc(
-					resource.TestCheckResourceAttr(name, "attack_type.0", "crlf"),
-					resource.TestCheckResourceAttr(name, "attack_type.1", "scanner"),
-					resource.TestCheckResourceAttr(name, "attack_type.2", "redir"),
-					resource.TestCheckResourceAttr(name, "attack_type.3", "ldapi"),
+					resource.TestCheckResourceAttr(name, "attack_type", "crlf"),
 					resource.TestCheckResourceAttr(name, "point.0.0", "get"),
 					resource.TestCheckResourceAttr(name, "point.0.1", "query"),
 					resource.TestCheckNoResourceAttr(name, "action"),
@@ -66,52 +55,10 @@ func TestAccRuleVpatchCreate_DefaultBranch(t *testing.T) {
 	})
 }
 
-func TestAccRuleVpatchCreate_FullSettings(t *testing.T) {
-	rnd := generateRandomResourceName(5)
-	name := "wallarm_rule_vpatch." + rnd
-	attackType := `"any", "sqli", "rce", "crlf", "nosqli", "ptrav", "xxe", "ptrav", "xss", "scanner", "redir", "ldapi"`
-	point := `["post"],["json_doc"],["array",0],["hash","password"]`
-	matchType := []string{"equal", "iequal", "regex", "absent"}
-	value := generateRandomResourceName(10) + ".example.com"
-
-	rule := vpatchTestingRule{
-		attackType: attackType,
-		point:      point,
-		matchType:  matchType,
-		value:      value,
-	}
-
-	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProviders,
-		CheckDestroy: testAccCheckWallarmRuleVpatchDestroy,
-		Steps: []resource.TestStep{
-			{
-				Config: testWallarmRuleVpatchFullSettingsConfig(rnd, rule),
-				Check: resource.ComposeTestCheckFunc(
-					resource.TestCheckResourceAttr(name, "attack_type.#", "12"),
-					resource.TestCheckResourceAttr(name, "action.#", "9"),
-					resource.TestCheckResourceAttr(name, "point.#", "4"),
-					resource.TestCheckResourceAttr(name, "point.0.#", "1"),
-					resource.TestCheckResourceAttr(name, "point.0.0", "post"),
-					resource.TestCheckResourceAttr(name, "point.1.#", "1"),
-					resource.TestCheckResourceAttr(name, "point.1.0", "json_doc"),
-					resource.TestCheckResourceAttr(name, "point.2.#", "2"),
-					resource.TestCheckResourceAttr(name, "point.2.0", "array"),
-					resource.TestCheckResourceAttr(name, "point.2.1", "0"),
-					resource.TestCheckResourceAttr(name, "point.3.#", "2"),
-					resource.TestCheckResourceAttr(name, "point.3.0", "hash"),
-					resource.TestCheckResourceAttr(name, "point.3.1", "password"),
-				),
-			},
-		},
-	})
-}
-
 func testWallarmRuleVpatchBasicConfig(resourceID, attackType, actionType, actionValue, actionPoint, point string) string {
 	return fmt.Sprintf(`
 resource "wallarm_rule_vpatch" "%[1]s" {
-  attack_type = ["%[2]s"]
+  attack_type = "%[2]s"
   action {
     type = "%[3]s"
     value = "%[4]s"
@@ -126,92 +73,9 @@ resource "wallarm_rule_vpatch" "%[1]s" {
 func testWallarmRuleVpatchDefaultBranchConfig(resourceID, attackType, point string) string {
 	return fmt.Sprintf(`
 resource "wallarm_rule_vpatch" "%[1]s" {
-	attack_type = [%[2]s]
+	attack_type = %[2]s
 	point = [%[3]s]
 }`, resourceID, attackType, point)
-}
-
-func testWallarmRuleVpatchFullSettingsConfig(resourceID string, rule vpatchTestingRule) string {
-	equal := rule.matchType[0]
-	iequal := rule.matchType[1]
-	regex := rule.matchType[2]
-	absent := rule.matchType[3]
-	return fmt.Sprintf(`
-resource "wallarm_rule_vpatch" "%[8]s" {
-
-	attack_type = [%[1]s]
-
-	action {
-		point = {
-		  instance = 1
-		}
-	}
-
-	action {
-		point = {
-		  instance = 1
-		}
-	}
-
-	action {
-		type = "%[3]s"
-		point = {
-		  action_name = "masking"
-		}
-	}
-
-	action {
-		type = "%[5]s"
-		point = {
-		  action_ext = ""
-		}
-	}
-
-	action {
-		type = "%[5]s"
-		point = {
-		  path = 0
-		}
-	}
-
-	action {
-		type = "%[3]s"
-		point = {
-		  method = "GET"
-		}
-	}
-
-	action {
-		type = "%[2]s"
-		point = {
-		  scheme = "https"
-		}
-	}
-
-	action {
-		type = "%[2]s"
-		point = {
-		  proto = "1.1"
-		}
-	}
-
-	action {
-		type = "%[4]s"
-		point = {
-		  uri = "/api/token[0-9A-Za-z]+"
-		}
-	}
-
-	action {
-		type = "%[3]s"
-		value = "%[6]s"
-		point = {
-		  header = "HOST"
-		}
-	}
-
-	point = [%[7]s]
-}`, rule.attackType, equal, iequal, regex, absent, rule.value, rule.point, resourceID)
 }
 
 func testAccCheckWallarmRuleVpatchDestroy(s *terraform.State) error {